+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Add breakpoint `condition` parameter by [@alexruperez](https://github.com/alexruperez).
+- Support Xcode Extension product type https://github.com/xcodeswift/xcproj/pull/190 by @briantkelley
 
 ## 1.7.0