+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 - Optimised performance of object lookups https://github.com/xcodeswift/xcproj/pull/191 by @kastiglione
+- **Breaking:** `PBXSourceTree` no longer has raw values and gained an assoicated value case to support custom locations https://github.com/xcodeswift/xcproj/pull/198 by @briantkelley
 ### Added
 - Add breakpoint `condition` parameter by [@alexruperez](https://github.com/alexruperez).